backend/requirements.txt

Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-pubsub==2.18.1
google-cloud-storage==2.12.0
google-cloud-firestore==2.13.1
google-cloud-secretmanager==2.18.1
google-cloud-documentai==2.22.0
sendgrid==6.10.0
PyPDF2==3.0.1
Jinja2==3.1.2
requests==2.31.0
playwright==1.40.0
gunicorn==21.2.0
